[PORT FROM MAIN]CLOTA: Rain of crashes
BZ: 33658

When a crash occurs it's reported by crashreport to crashtool server.
When a rain of crashes (following of crashes which have the same signature
(type,data0,data1,data2) occurs, all crashes are reported and it could be an
important process cost and bandwidth consumption which is not usefull. It could
also reduce the bitrate authorization if the "reporter" uses his 3G connexion
to report the crashes.
Moreover, it's not very interesting to report all crashes of a rain of
crashes. It's wiser to report the first crashes of the rain and to report the
number of crashes in the rain.

The goal of this patch is to detect the beginning of a rain and to report only
the first ten crashes. Then, at each fibonnacci sequence elements (13,21...),
rain of crashes events are reported. They indicate the crash's signature of the
rain and the number of crashes present in the rain.Only JAVACRASH,ANR and
TOMBSTONE's rain of crashes are managed.
Moreover this patch solves also the duplicate ANRs problem.

When a crash occurs, PhoneDoctor checks if it's a JAVACRASH,ANR or TOMBSTONE.
In case of JAVACRASH,ANR,TOMBSTONE PhoneDoctor checks if the crash is a part
of a rain of crashes or not.
A new table rain_of_crashes was created. This table stores all rain of crashes.
It's compound with the signature of the crash(type,data0,data1,data2), the date
of the last apperance of the crash, the number of crashes present in the rain,
the eventId of the first crash of the rain and the fibonnacci sequence current
and last number.

PhoneDoctor checks if a rain of crashes is running for this crash or not.
If it's the case, PhoneDoctor checks that the rain is still running(last crash
of the rain occured less than 10 minutes ago). If it's the case rain_of_crashes
is updated for this crash (occurences and date are changed). If occurences is
equal to the next fibonacci sequence number, a new RAIN event is created.
If the rain is finished, PhoneDoctor checks that the crash is in a new rain or
not (check that almost 10 crashes with the same signature has occured since
last crash and during the last hour). If it's the case a RAIN event is created
if there are crashes not reported in the last rain.
(occurences not equal to a fibonacci sequence number). Then rain_of_crashes is
updated for this crash signature and a new RAIN event is created.
In other cases, the crash is not a part of a rain of crashes and it's reported
like the other crashes.

All events which are in a rain of crashes are black listed and stored in
black_events table.(eventId of the event, crashdir, eventId associated with
the rain, reason of the black list (FAKE or DUPLICATE)).

In case of ANR, the duplicate ANR are checked. When an ANR occurs, PhoneDoctor
checks that the TraceId File associated with this ANR is not aleady present in
an old ANR. If it's the case the ANR is black listed with the reason FAKE.

Each time an event occurs PhoneDoctor checks that a rain of crashes has not
expired (last crash occured in the last hour). If it's the case,expired rain(s)
are deleted from the rain_of_crashes table,and crashfile of black-listed events
associated with those rains are modified by adding _FAKE or _DUPLICATE at
the end of the crash type.

Update: test with chek event during upload.
        add the parse of system_app_anr*.gz

Reviewed-on: http://android.intel.com:8080/73604
Tested-by: Guerard, Christophe <christophe.guerard@intel.com>
Tested-by: buildbot <buildbot@intel.com>
Reviewed-on: http://android.intel.com:8080/78515
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Tested-by: cactus <cactus@intel.com>

+++ crashreport/src/com/intel/crashreport/CrashReport.java
@@ -161,6 +161,7 @@ public class CrashReport extends Application {
 		String myBuild;
 		Event event;
 		NotificationMgr nMgr;
+		BlackLister blackLister = new BlackLister();
 
 		db = new EventDB(getApplicationContext());
 		myBuild = ((CrashReport) getApplicationContext()).getMyBuild().toString();
@@ -168,6 +169,7 @@ public class CrashReport extends Application {
 
 		try {
 			db.open();
+			blackLister.setDb(db);
 			histFile = new HistoryEventFile();
 
 			while (histFile.hasNext()) {
@@ -175,32 +177,35 @@ public class CrashReport extends Application {
 				if (histEventLine.length() != 0) {
 					HistoryEvent histEvent = new HistoryEvent(histEventLine);
 					if (histEvent.getEventId().replaceAll("0", "").length() != 0) {
-						if (!db.isEventInDb(histEvent.getEventId())) {
+						if (!db.isEventInDb(histEvent.getEventId()) && !db.isEventInBlackList(histEvent.getEventId())) {
 							event = new Event(histEvent, myBuild);
-							long ret = db.addEvent(event);
-							if (ret == -1)
-								Log.w(from+": Event error when added to DB, " + event.toString());
-							else if (ret == -2)
-								Log.w(from+": Event name " +histEvent.getEventName() + " unkown, addition in DB canceled");
-							else if (ret == -3)
-								Log.w(from+": Event " +event.toString() + " with wrong date, addition in DB canceled");
-							else {
-								if (event.getEventName().contentEquals("REBOOT")) {
-									if (event.getType().contentEquals("SWUPDATE")){
-										db.deleteEventsBeforeUpdate(event.getEventId());
-									}else{
-										db.updateEventsNotReadyBeforeREBOOT(event.getEventId());
+							blackLister.cleanRain(event.getDate());
+							if (!blackLister.blackList(event)) {
+								long ret = db.addEvent(event);
+								if (ret == -1)
+									Log.w(from+": Event error when added to DB, " + event.toString());
+								else if (ret == -2)
+									Log.w(from+": Event name " +histEvent.getEventName() + " unkown, addition in DB canceled");
+								else if (ret == -3)
+									Log.w(from+": Event " +event.toString() + " with wrong date, addition in DB canceled");
+								else {
+									if (event.getEventName().contentEquals("REBOOT")) {
+										if (event.getType().contentEquals("SWUPDATE")){
+											db.deleteEventsBeforeUpdate(event.getEventId());
+										}else{
+											db.updateEventsNotReadyBeforeREBOOT(event.getEventId());
+										}
 									}
-								}
-								if (event.getEventName().equals("BZ")) {
-									BZFile bzfile = new BZFile(event.getCrashDir());
-									db.addBZ(event.getEventId(), bzfile, event.getDate());
-									Log.d(from+": BZ added in DB, " + histEvent.getEventId());
-								}
-								Log.d(from+": Event successfully added to DB, " + event.toString());
-								if (!event.isDataReady()) {
-									Timer timer = new Timer(true);
-									timer.schedule(new NotifyCrashTask(event.getEventId(),getApplicationContext()), CRASH_POSTPONE_DELAY*1000);
+									if (event.getEventName().equals("BZ")) {
+										BZFile bzfile = new BZFile(event.getCrashDir());
+										db.addBZ(event.getEventId(), bzfile, event.getDate());
+										Log.d(from+": BZ added in DB, " + histEvent.getEventId());
+									}
+									Log.d(from+": Event successfully added to DB, " + event.toString());
+									if (!event.isDataReady()) {
+										Timer timer = new Timer(true);
+										timer.schedule(new NotifyCrashTask(event.getEventId(),getApplicationContext()), CRASH_POSTPONE_DELAY*1000);
+									}
 								}
 							}
 						} else
